pyproject.toml
[project]
name = "ghost-home-assistant"
version = "0.2.0"
description = "Ghost integration for Home Assistant"
readme = "README.md"
license = "MIT"
requires-python = ">=3.12"

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.0.0",
    "pytest-homeassistant-custom-component>=0.13.0",
    "aioghost>=0.2.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
pythonpath = ["."]

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "C4", "W"]
